requirements.txt
uvicorn~=0.34.3
fastapi~=0.115.13
passlib~=1.7.4
PyJWT~=2.10.1
python-multipart~=0.0.20
bcrypt==4.0.1
psycopg2-binary~=2.9.10